Remove duplicate check (#576)
* Remove duplicate check

1. It was not an error, just log spam.
2. It was the same for EVERY package.

If we decide this is an error, if should be package-specific. But there
would need to be a way to override it, because lots of packages have
legitimate duplicates. For example, all of lodash has the same project
url. And multiple versions of the same project have the same name.

* Remove surrounding code too